code cleanup && fix
