+++ SuperTwin/pcp_rt7-laptop-vivo_monitor.conf
@@ -1,0 +1,32 @@
+[options]
+influx_server = http://127.0.0.1:8086
+influx_db = rt7-laptop-vivo
+influx_tags = tag=_monitor
+source = localhost
+
+
+[configured]
+kernel.all.pressure.cpu.some.total = ,,
+hinv.cpu.clock = ,,
+lmsensors.coretemp_isa_0000.package_id_0 = ,,
+kernel.pernode.cpu.idle = ,,
+kernel.percpu.cpu.idle = ,,
+disk.dev.read = ,,
+disk.dev.write = ,,
+disk.dev.total = ,,
+disk.dev.read_bytes = ,,
+disk.dev.write_bytes = ,,
+disk.dev.total_bytes = ,,
+swap.pagesin = ,,
+kernel.all.nusers = ,,
+kernel.all.nprocs = ,,
+network.all.in.bytes = ,,
+network.all.out.bytes = ,,
+mem.util.used = ,,
+mem.util.free = ,,
+mem_use = mem_use 
+mem_use.formula = proc.psinfo.rss / 1024 
+mem_use.label = "mem_use"
+cpu_use = cpu_use 
+cpu_use.formula = 0.1 * (proc.psinfo.utime + proc.psinfo.stime) 
+cpu_use.label = "cpu_use"
