+++ SuperTwin/perfevent.conf
@@ -1,28 +1,1 @@
-[ivb_ep]
-UNHALTED_CORE_CYCLES
-UNHALTED_REFERENCE_CYCLES
-INSTRUCTION_RETIRED
-LLC_REFERENCES
-LLC_MISSES
-BRANCH_INSTRUCTION_RETIRED
-MISPREDICTED_BRANCH_RETIRED
-BACLEARS:ANY
-L1D:REPLACEMENT
-L2_LINES_IN:ALL
-L2_TRANS:L2_WB
-MEM_LOAD_RETIRED:L1_HIT
-MEM_LOAD_RETIRED:L1_MISS
-UOPS_DISPATCHED_PORT:PORT_0
-UOPS_DISPATCHED_PORT:PORT_1
-UOPS_DISPATCHED_PORT:PORT_5
-FP_COMP_OPS_EXE:X87
-LOAD_HIT_PRE:HW_PF
-FP_COMP_OPS_EXE:SSE_PACKED_SINGLE
-FP_COMP_OPS_EXE:SSE_PACKED_DOUBLE
-SIMD_FP_256:PACKED_SINGLE
-SIMD_FP_256:PACKED_DOUBLE
-MEM_UOPS_RETIRED:ALL_LOADS
-MEM_UOPS_RETIRED:ALL_STORES
-UOPS_ISSUED:STALL_CYCLES
-RAPL_ENERGY_PKG
-RAPL_ENERGY_DRAM
+[icl]
